// app.client/src/api/models/FormSubmissionListVM.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormSubmissionListVM = {
    formSubmissionId?: string | null;
    formDataId?: string | null;
    alternateFormId?: string | null;
    jsonData?: string | null;
    isVoided?: boolean;
    createdBy?: string | null;
    createdDate?: string;
    lastModifiedBy?: string | null;
    lastModifiedDate?: string | null;
};